+++ Idle_Game/app/src/main/java/com/example/idle_game/di/RepoModule.kt
@@ -72,6 +72,6 @@ class RepoModule {
     fun providesDatabase(@ApplicationContext context: Context): GameDatabase {
         return Room.databaseBuilder(
             context, GameDatabase::class.java, "gamedatabase"
-        ).build()
+        ).fallbackToDestructiveMigration().build()
     }
 }
