+++ Idle_Game/app/src/main/java/com/example/idle_game/ui/theme/Color.kt
@@ -38,10 +38,6 @@ val surfaceContainerLight = Color(0xFFF0EEE6)
 val surfaceContainerHighLight = Color(0xFFEAE8E1)
 val surfaceContainerHighestLight = Color(0xFFE4E2DB)
 
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-val bitcoinBackground = Color(0xFFF7931A)
 val primaryLightMediumContrast = Color(0xFF3B4716)
 val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
 val primaryContainerLightMediumContrast = Color(0xFF6C7A43)
