+++ Idle_Game/app/src/main/java/com/example/idle_game/ui/views/models/InventoryViewModel.kt
@@ -1,6 +1,8 @@
 package com.example.idle_game.ui.views.models
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.idle_game.data.database.models.InventoryData
+import com.example.idle_game.data.database.models.ShopData
 import com.example.idle_game.data.repositories.GameRepository
 import com.example.idle_game.ui.views.states.InventoryViewState
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -8,6 +10,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.first
 
 @HiltViewModel
 class InventoryViewModel @Inject constructor(
@@ -25,6 +28,23 @@ class InventoryViewModel @Inject constructor(
             gameRepository.updateShop()
             _uiStateFlow.value = _uiStateFlow.value.copy(shopData = shopData)
             _uiStateFlow.value = _uiStateFlow.value.copy(inventoryData = inventoryData)
+        }
+    }
+
+    //Get amount of certain items
+    fun getAmountOfItems(item: ShopData, inventoryData: InventoryData): Int {
+        return when (item.name) {
+            "low Boost" -> inventoryData.lowBoosts
+            "medium Boost" -> inventoryData.mediumBoosts
+            "high Boost" -> inventoryData.highBoosts
+            "low passive" -> inventoryData.hackersLvl1 + inventoryData.hackersLvl2 + inventoryData.hackersLvl3 + inventoryData.hackersLvl4 + inventoryData.hackersLvl5
+            "medium passive" -> inventoryData.cryptoMinersLvl1 + inventoryData.cryptoMinersLvl2 + inventoryData.cryptoMinersLvl3 + inventoryData.cryptoMinersLvl4 + inventoryData.cryptoMinersLvl5
+            "high passive" -> inventoryData.botnetsLvl1 + inventoryData.botnetsLvl2 + inventoryData.botnetsLvl3 + inventoryData.botnetsLvl4 + inventoryData.botnetsLvl5
+            "upgrade lvl 2" -> inventoryData.upgradeLvl2
+            "upgrade lvl 3" -> inventoryData.upgradeLvl3
+            "upgrade lvl 4" -> inventoryData.upgradeLvl4
+            "upgrade lvl 5" -> inventoryData.upgradeLvl5
+            else -> 0
         }
     }
 }
