+++ Idle_Game/app/src/main/java/com/example/idle_game/ui/views/states/InventoryViewState.kt
@@ -3,7 +3,6 @@ package com.example.idle_game.ui.views.states
 import com.example.idle_game.data.database.models.InventoryData
 import com.example.idle_game.data.database.models.ShopData
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.emptyFlow
 
 data class InventoryViewState(
